Fixed 500 empty body responses failing import.
Some Steam apps have started returning permanent HTTP 500 errors. We do
not have the tooling to heuristically detect whether they're permanent
at present so, for now, this case is assumed to always be fatal.

Once the Retry library passes the retry count to the callback and the
exception handler stops sleeping the thread, we can revisit this.

Some app IDs that throw 500s: 5648, 5688, 5689.